+++ sdk/src/wallet/account/mod.rs
@@ -367,7 +367,9 @@ impl AccountInner {
                     }
                 }
 
-                filtered_outputs.push(output.clone());
+                if filter.alias_ids.is_none() && filter.foundry_ids.is_none() && filter.nft_ids.is_none() {
+                    filtered_outputs.push(output.clone());
+                }
             }
 
             Ok(filtered_outputs)