+++ sdk/CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Call `ledger.set_non_interactive_mode()` only if it's a debug app;
+- Don't return other output types if alias/nft/foundry ids are provided in the `FilterOptions` for `Account::{outputs(), unspent_outputs()`;
 
 ## 0.4.0 - 2023-07-14